requirements.txt
llama-index
chromadb
unstructured[pdf]
pypdf
python-dotenv
streamlit
openai
rank-bm25
